Moved most of the jacrev test code logic to common test file.
